updated validation

+++ ios/batm/UI/P2PFlow/Edit Trade/P2PEditTradeViewController.swift
@@ -7,8 +7,8 @@ import MaterialComponents
 struct P2PEditTradeDataModel: Encodable {
   let id: String
   let price: Double
-  let minLimit: Int
-  let maxLimit: Int
+  let minLimit: Double
+  let maxLimit: Double
   let paymentMethods: String
   let terms: String
   
@@ -24,8 +24,8 @@ protocol P2PEditTradeViewControllerDelegate: AnyObject {
 class P2PEditTradeViewController: UIViewController {
   
   let selectedType: P2PSellBuyViewType
-  var minRange: Int?
-  var maxRange: Int?
+  var minRange: Double?
+  var maxRange: Double?
   
   let currentModel: P2PEditTradeDataModel
